+++ samples/Calculator/fabric/android/app/build.gradle
@@ -77,9 +77,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.calculatorexperimental"
+    namespace "com.calculatorfabric"
     defaultConfig {
-        applicationId "com.calculatorexperimental"
+        applicationId "com.calculatorfabric"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
