+++ samples/Calculator/fabric/android/app/src/main/java/com/calculatorfabric/MainApplication.kt
@@ -1,4 +1,4 @@
-package com.calculatorexperimental
+package com.calculatorfabric
 
 import android.app.Application
 import com.facebook.react.PackageList
