+++ samples/Calculator/fabric/android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">CalculatorExperimental</string>
+    <string name="app_name">CalculatorFabric</string>
 </resources>
