+++ samples/Calculator/fabric/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "CalculatorExperimental",
-  "displayName": "CalculatorExperimental"
+  "name": "CalculatorFabric",
+  "displayName": "CalculatorFabric"
 }
