+++ samples/Calculator/fabric/ios/CalculatorFabric.xcodeproj/xcshareddata/xcschemes/CalculatorFabric.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "CalculatorExperimental.app"
-               BlueprintName = "CalculatorExperimental"
-               ReferencedContainer = "container:CalculatorExperimental.xcodeproj">
+               BuildableName = "CalculatorFabric.app"
+               BlueprintName = "CalculatorFabric"
+               ReferencedContainer = "container:CalculatorFabric.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "CalculatorExperimentalTests.xctest"
-               BlueprintName = "CalculatorExperimentalTests"
-               ReferencedContainer = "container:CalculatorExperimental.xcodeproj">
+               BuildableName = "CalculatorFabricTests.xctest"
+               BlueprintName = "CalculatorFabricTests"
+               ReferencedContainer = "container:CalculatorFabric.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "CalculatorExperimental.app"
-            BlueprintName = "CalculatorExperimental"
-            ReferencedContainer = "container:CalculatorExperimental.xcodeproj">
+            BuildableName = "CalculatorFabric.app"
+            BlueprintName = "CalculatorFabric"
+            ReferencedContainer = "container:CalculatorFabric.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "CalculatorExperimental.app"
-            BlueprintName = "CalculatorExperimental"
-            ReferencedContainer = "container:CalculatorExperimental.xcodeproj">
+            BuildableName = "CalculatorFabric.app"
+            BlueprintName = "CalculatorFabric"
+            ReferencedContainer = "container:CalculatorFabric.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
