+++ samples/Calculator/fabric/ios/CalculatorFabric/Images.xcassets/AppIcon.appiconset/Contents.json
