+++ samples/Calculator/fabric/ios/CalculatorFabric/Images.xcassets/Contents.json
