+++ samples/Calculator/fabric/ios/CalculatorFabric/PrivacyInfo.xcprivacy
