+++ samples/Calculator/fabric/ios/CalculatorFabricTests/Info.plist
