+++ samples/Calculator/fabric/ios/Podfile
@@ -14,7 +14,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'CalculatorExperimental' do
+target 'CalculatorFabric' do
   config = use_native_modules!
 
   use_react_native!(
@@ -23,7 +23,7 @@ target 'CalculatorExperimental' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'CalculatorExperimentalTests' do
+  target 'CalculatorFabricTests' do
     inherit! :complete
     # Pods for testing
   end
