+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Images/LockScreenLogo.scale-200.png
