+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Images/SplashScreen.scale-200.png
