+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Images/Square150x150Logo.scale-200.png
