+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Images/Square44x44Logo.scale-200.png
