+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Images/Square44x44Logo.targetsize-24_altform-unplated.png
