+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Images/StoreLogo.png
