+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Images/Wide310x150Logo.scale-200.png
