+++ samples/Calculator/fabric/windows/CalculatorFabric.Package/Package.appxmanifest
@@ -7,13 +7,13 @@
   IgnorableNamespaces="uap rescap">
 
   <Identity
-    Name="CalculatorExperimental"
-    Publisher="CN=React Native Windows Sample"
+    Name="CalculatorFabric"
+    Publisher="CN=tatianakapos"
     Version="1.0.0.0" />
 
   <Properties>
-    <DisplayName>CalculatorExperimental</DisplayName>
-    <PublisherDisplayName>CN=React Native Windows Sample</PublisherDisplayName>
+    <DisplayName>CalculatorFabric</DisplayName>
+    <PublisherDisplayName>tatianakapos</PublisherDisplayName>
     <Logo>Images\StoreLogo.png</Logo>
   </Properties>
 
@@ -31,8 +31,8 @@
       Executable="$targetnametoken$.exe"
       EntryPoint="$targetentrypoint$">
       <uap:VisualElements
-        DisplayName="CalculatorExperimental"
-        Description="CalculatorExperimental"
+        DisplayName="CalculatorFabric"
+        Description="CalculatorFabric"
         BackgroundColor="transparent"
         Square150x150Logo="Images\Square150x150Logo.png"
         Square44x44Logo="Images\Square44x44Logo.png">
