+++ samples/Calculator/fabric/windows/CalculatorFabric.sln
@@ -1,0 +1,63 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32929.385
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5}") = "CalculatorFabric.Package", "CalculatorFabric.Package\CalculatorFabric.Package.wapproj", "{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "CalculatorFabric", "CalculatorFabric\CalculatorFabric.vcxproj", "{1EE64A59-709D-44C9-AD25-589DFF90ECCA}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Debug|ARM64 = Debug|ARM64
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+		Release|ARM64 = Release|ARM64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|x64.ActiveCfg = Debug|x64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|x64.Build.0 = Debug|x64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|x64.Deploy.0 = Debug|x64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|x86.ActiveCfg = Debug|x86
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|x86.Build.0 = Debug|x86
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|x86.Deploy.0 = Debug|x86
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|ARM64.Build.0 = Debug|ARM64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Debug|ARM64.Deploy.0 = Debug|ARM64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|x64.ActiveCfg = Release|x64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|x64.Build.0 = Release|x64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|x64.Deploy.0 = Release|x64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|x86.ActiveCfg = Release|x86
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|x86.Build.0 = Release|x86
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|x86.Deploy.0 = Release|x86
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|ARM64.ActiveCfg = Release|ARM64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|ARM64.Build.0 = Release|ARM64
+		{E5E661AB-57CA-431F-9C71-8CD37E1E88FA}.Release|ARM64.Deploy.0 = Release|ARM64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|x64.ActiveCfg = Debug|x64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|x64.Build.0 = Debug|x64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|x64.Deploy.0 = Debug|x64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|x86.ActiveCfg = Debug|Win32
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|x86.Build.0 = Debug|Win32
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|x86.Deploy.0 = Debug|Win32
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|ARM64.ActiveCfg = Debug|ARM64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|ARM64.Build.0 = Debug|ARM64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Debug|ARM64.Deploy.0 = Debug|ARM64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|x64.ActiveCfg = Release|x64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|x64.Build.0 = Release|x64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|x64.Deploy.0 = Release|x64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|x86.ActiveCfg = Release|Win32
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|x86.Build.0 = Release|Win32
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|x86.Deploy.0 = Release|Win32
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|ARM64.ActiveCfg = Release|ARM64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|ARM64.Build.0 = Release|ARM64
+		{1EE64A59-709D-44C9-AD25-589DFF90ECCA}.Release|ARM64.Deploy.0 = Release|ARM64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D43FAD39-F619-437D-BB40-04A3982ACB6A}
+	EndGlobalSection
+EndGlobal
