+++ samples/Calculator/fabric/windows/CalculatorFabric/.gitignore
