+++ samples/Calculator/fabric/windows/CalculatorFabric/AutolinkedNativeModules.g.props
