+++ samples/Calculator/fabric/windows/CalculatorFabric/CalculatorFabric.ico
