+++ samples/Calculator/fabric/windows/CalculatorFabric/CalculatorFabric.rc
