+++ samples/Calculator/fabric/windows/CalculatorFabric/resource.h
@@ -1,6 +1,6 @@
 //
 // Microsoft Visual C++ generated include file.
-// Used by CalculatorExperimental.rc
+// Used by CalculatorFabric.rc
 
 #define IDI_ICON1 1008
 // Next default values for new objects
