+++ samples/Calculator/fabric/windows/CalculatorFabric/small.ico
